+++ tests/testthat/test-tm_t_events_by_grade.R
@@ -65,9 +65,11 @@ test_that("template_events_by_grade generates standard expressions", {
         alt_counts_df = adsl
       )
     ),
-    pruned_and_sorted_result = quote({
-      result <- result %>%
-        trim_rows() %>%
+    prune = quote({
+      pruned_result <- result %>% trim_rows()
+    }),
+    sort = quote({
+      pruned_and_sorted_result <- pruned_result %>%
         sort_at_path(
           path = "AEBODSYS",
           scorefun = cont_n_onecol(length(levels(adsl$ACTARM)) + 1),
@@ -77,7 +79,106 @@ test_that("template_events_by_grade generates standard expressions", {
           scorefun = cont_n_onecol(length(levels(adsl$ACTARM)) + 1),
           decreasing = TRUE
         )
-      result
+      pruned_and_sorted_result
+    })
+  )
+
+  expect_equal(result, expected)
+})
+
+test_that("template_events_by_grade generates standard expressions with pruning conditions", {
+
+  result <- template_events_by_grade(
+    dataname = "adae",
+    parentname = "adsl",
+    arm_var = "ACTARM",
+    hlt = "AEBODSYS",
+    llt = "AEDECOD",
+    grade = "AESEV",
+    prune_freq = 0.4,
+    prune_diff = 0.1,
+    add_total = TRUE,
+    drop_arm_levels = TRUE
+  )
+
+  expected <- list(
+    data = quote({
+      anl <- adae # nolintr
+      anl <- anl %>% mutate(ACTARM = droplevels(ACTARM))
+      arm_levels <- levels(anl[["ACTARM"]])
+      adsl <- adsl %>% filter(ACTARM %in% arm_levels)
+      adsl <- adsl %>% mutate(ACTARM = droplevels(ACTARM))
+      anl <- df_explicit_na(anl, na_level = "")
+      adsl <- df_explicit_na(adsl, na_level = "")
+      grade_groups <- list("- Any Intensity -" = levels(adae$AESEV))
+    }),
+    layout_prep = quote(split_fun <- drop_split_levels),
+    layout = quote(
+      lyt <- basic_table() %>%
+        split_cols_by("ACTARM") %>%
+        add_overall_col(label = "All Patients") %>%
+        add_colcounts() %>%
+        summarize_occurrences_by_grade(
+          var = "AESEV",
+          grade_groups = grade_groups) %>%
+        split_rows_by(
+          "AEBODSYS",
+          child_labels = "visible",
+          nested = TRUE,
+          indent_mod = -1L,
+          split_fun = split_fun,
+          label_pos = "topleft",
+          split_label = var_labels(adae["AEBODSYS"], fill = TRUE)
+        ) %>%
+        summarize_occurrences_by_grade(
+          var = "AESEV",
+          grade_groups = grade_groups) %>%
+        split_rows_by(
+          "AEDECOD",
+          child_labels = "visible",
+          nested = TRUE,
+          indent_mod = -1L,
+          split_fun = split_fun,
+          label_pos = "topleft",
+          split_label = var_labels(adae["AEDECOD"], fill = TRUE)
+        ) %>%
+        summarize_num_patients(
+          var = "USUBJID",
+          .stats = "unique",
+          .labels = c("- Any Intensity -")) %>%
+        count_occurrences_by_grade(var = "AESEV", .indent_mods = -1L) %>%
+        append_varlabels(adae, "AESEV", indent = 2L)
+    ),
+    table = quote(
+      result <- build_table(
+        lyt = lyt,
+        df = anl,
+        alt_counts_df = adsl
+      )
+    ),
+    prune = quote({
+      pruned_result <- result %>% trim_rows()
+      col_indices <- 1:(ncol(result) - TRUE)
+      row_condition <- has_fraction_in_any_col(atleast = 0.4, col_indices = col_indices) &
+        has_fractions_difference(atleast = 0.1, col_indices = col_indices)
+      pruned_result <- pruned_result %>% prune_table(keep_rows(row_condition))
+    }),
+    sort = quote({
+      pruned_and_sorted_result <- pruned_result %>%
+        sort_at_path(
+          path = "AEBODSYS",
+          scorefun = cont_n_onecol(length(levels(adsl$ACTARM)) + 1),
+          decreasing = TRUE) %>%
+        sort_at_path(
+          path = c("AEBODSYS", "*", "AEDECOD"),
+          scorefun = cont_n_onecol(length(levels(adsl$ACTARM)) + 1),
+          decreasing = TRUE
+        )
+      criteria_fun <- function(tr) {
+        is(tr, "ContentRow")
+      }
+      pruned_and_sorted_result <- trim_rows(pruned_and_sorted_result, criteria = criteria_fun)
+      pruned_and_sorted_result
     })
   )
 
@@ -151,9 +252,11 @@ test_that("template_events_by_grade without adding total column option works as 
         alt_counts_df = adsl
       )
     ),
-    pruned_and_sorted_result = quote({
-      result <- result %>%
-        trim_rows() %>%
+    prune = quote({
+      pruned_result <- result %>% trim_rows()
+    }),
+    sort = quote({
+      pruned_and_sorted_result <- pruned_result %>%
         sort_at_path(
           path = "AEBODSYS",
           scorefun = cont_n_allcols,
@@ -164,7 +267,7 @@ test_that("template_events_by_grade without adding total column option works as 
           scorefun = cont_n_allcols,
           decreasing = TRUE
         )
-      result
+      pruned_and_sorted_result
     })
   )
 
@@ -226,15 +329,17 @@ test_that("template_events_by_grade with hlt only works", {
         alt_counts_df = adsl
       )
     ),
-    pruned_and_sorted_result = quote({
-      result <- result %>%
-        trim_rows() %>%
+    prune = quote({
+      pruned_result <- result %>% trim_rows()
+    }),
+    sort = quote({
+      pruned_and_sorted_result <- pruned_result %>%
         sort_at_path(
           path = "AEBODSYS",
           scorefun = cont_n_onecol(length(levels(adsl$ACTARM)) + 1),
           decreasing = TRUE
         )
-      result
+      pruned_and_sorted_result
     })
   )
 
@@ -345,8 +450,7 @@ test_that("template_events_col_by_grade generates standard expressions", {
       pruned_and_sorted_result <- sorted_result %>%
         trim_rows(criteria = criteria_fun) %>%
         prune_table(keep_rows(at_least_percent_any))
-      result <- pruned_and_sorted_result
-      result
+      pruned_and_sorted_result
     })
   )
 
@@ -448,8 +552,7 @@ test_that("template_events_col_by_grade generates STREAM variant 8", {
       pruned_and_sorted_result <- sorted_result %>%
         trim_rows(criteria = criteria_fun) %>%
         prune_table(keep_rows(at_least_percent_any))
-      result <- pruned_and_sorted_result
-      result
+      pruned_and_sorted_result
     })
   )
 
@@ -551,8 +654,7 @@ test_that("template_events_col_by_grade without adding total column option works
       pruned_and_sorted_result <- sorted_result %>%
         trim_rows(criteria = criteria_fun) %>%
         prune_table(keep_rows(at_least_percent_any))
-      result <- pruned_and_sorted_result
-      result
+      pruned_and_sorted_result
     })
   )
 
@@ -653,8 +755,7 @@ test_that("template_events_col_by_grade without dropping arm levels option works
       pruned_and_sorted_result <- sorted_result %>%
         trim_rows(criteria = criteria_fun) %>%
         prune_table(keep_rows(at_least_percent_any))
-      result <- pruned_and_sorted_result
-      result
+      pruned_and_sorted_result
     })
   )